+++ lib/QECCore/ext/QECCoreNemoExt.jl
@@ -6,18 +6,14 @@ using DocStringExtensions
 import Nemo
 import Nemo: GF, matrix, rank, transpose
 
-import QECCore: TillichZemor, code_k, _create_circulant_matrix, _create_matrix_M_deterministic
+import QECCore: code_k, parity_matrix_x, parity_matrix_z
 
-function QECCore.code_k(c::TillichZemor)
-    C = _create_circulant_matrix(c.m)
-    M = _create_matrix_M_deterministic(c.m, c.n, c.r)
-    H = hcat(C, M)
-    H_gf2 = matrix(GF(2), H)
-    Ht_gf2 = transpose(H_gf2)
-    k = c.n - rank(H_gf2)
-    kT = c.m - rank(Ht_gf2)  # c.m == size(H, 1)
-    k_q = k^2 + kT^2
-    return k_q
+function QECCore.code_k(c::AbstractQECC)
+    n = code_n(c)
+    F₂ = GF(2)
+    rank_Hx = rank(matrix(F₂, parity_matrix_x(c)))
+    rank_Hz = rank(matrix(F₂, parity_matrix_z(c)))
+    return n - rank_Hx - rank_Hz
 end
 
 end # module